Convert `mastodon/api` to Typescript
